move IngestionParams to internal package
