+++ dao-api/src/main/pegasus/com/linkedin/metadata/internal/IngestionParams.pdl
@@ -1,4 +1,6 @@
-namespace com.linkedin.metadata.events
+namespace com.linkedin.metadata.internal
+
+import com.linkedin.metadata.events.IngestionMode
 
 /**
  * Record defining ingestion-related parameters that can be passed into DAO API calls.
