+++ dao-impl/ebean-dao/src/main/java/com/linkedin/metadata/dao/ImmutableLocalDAO.java
@@ -8,8 +8,8 @@ import com.linkedin.data.template.RecordTemplate;
 import com.linkedin.data.template.UnionTemplate;
 import com.linkedin.metadata.dao.producer.DummyMetadataEventProducer;
 import com.linkedin.metadata.dao.utils.RecordUtils;
-import com.linkedin.metadata.events.IngestionParams;
 import com.linkedin.metadata.events.IngestionTrackingContext;
+import com.linkedin.metadata.internal.IngestionParams;
 import io.ebean.Ebean;
 import java.io.IOException;
 import java.io.InputStream;
