+++ restli-resources/src/main/java/com/linkedin/metadata/restli/BaseAspectV2Resource.java
@@ -9,8 +9,8 @@ import com.linkedin.metadata.dao.AspectKey;
 import com.linkedin.metadata.dao.BaseLocalDAO;
 import com.linkedin.metadata.dao.ListResult;
 import com.linkedin.metadata.dao.utils.ModelUtils;
-import com.linkedin.metadata.events.IngestionParams;
 import com.linkedin.metadata.events.IngestionTrackingContext;
+import com.linkedin.metadata.internal.IngestionParams;
 import com.linkedin.metadata.query.ListResultMetadata;
 import com.linkedin.metadata.validator.ValidationUtils;
 import com.linkedin.parseq.Task;
